Added wrapper for text writers. Finished the UI on 99%.
Some bugs with smart pointers still persist. UI is almost done, will wrap it in menu class next time. I added wrappers for input and output files, that will manage the stream resources. Still need to the the main factories, but they should be a piece of cake after the design of the Menu, together with the base hierarchy design are cleared. Unit tests still are to be done