Styles equalities, Composite start
I have added a functionality of equality comparison between transformations. This is later used by the dynamic addition/ removal of styles which is implemented as part of part 4 of the task. Covered the textTransformation with additional unit tests. Composite pattern is hinted as people can add many styles on top of the simple transformation