Added visitors
Added 2 visitors, which will calculate the size of a file/directory or implement hashing of a file. The end product of the program is complete, by specifying a filename a person can extract the info(hashes/ sizes) of a subtree and write them to a resulting file. Cosmetic upgrades still left to do : 1) Observer of the hashing process
2) Memento which can be used to pause the hashing process